refactor(web): 优化代码质量和错误处理
- 将多个函数中的 if let Err 处理改为 ? 操作符简化错误传播
- 修复数据绑定中的类型转换问题，移除不必要的 as i64 类型转换
- 优化条件检查逻辑，将 if let Some(_) 替换为 is_some() 方法
- 在模板分析服务中将正则表达式提升为 LazyLock 静态编译以提高性能
- 调整文件路径处理逻辑，使用 strip_prefix 的结果验证替代 unwrap_or
- 重新排列模块导入顺序以符合代码风格规范
- 修复 PagedResponse 构造中的类型参数传递问题
- 移除代码中的冗余格式化操作和无用的类型转换

+++ crates/repositories/src/template_repository.rs
@@ -33,7 +33,7 @@ impl TemplateRepository {
         .bind(template_id)
         .bind(&request.name)
         .bind(&request.description)
-        .bind(request.category_id as i64)
+        .bind(request.category_id)
         .bind(&request.template_type)
         .bind(&request.type_config)
         .bind(&request.introduction)
@@ -219,7 +219,7 @@ impl TemplateRepository {
             );
 
             let templates = sqlx::query_as::<_, Template>(&sql)
-                .bind(category_id as i64)
+                .bind(category_id)
                 .bind(page_size as i64)
                 .bind(offset as i64)
                 .fetch_all(&self.pool)
@@ -230,7 +230,7 @@ impl TemplateRepository {
                 vis_filter
             );
             let total: i64 = sqlx::query_scalar(&count_sql)
-                .bind(category_id as i64)
+                .bind(category_id)
                 .fetch_one(&self.pool)
                 .await
                 .unwrap_or(0);
@@ -295,12 +295,7 @@ impl TemplateRepository {
             (templates, total)
         };
 
-        Ok(PagedResponse::new(
-            templates,
-            total as u32,
-            page as u32,
-            page_size as u32,
-        ))
+        Ok(PagedResponse::new(templates, total as u32, page, page_size))
     }
 
     /// 切换推荐状态
